requirements.txt
asgiref==3.2.10
Django==3.1.2
django-widget-tweaks==1.4.8
Pillow==8.0.1
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
gunicorn==20.0.4
django-import-export==2.4.0
dj-database-url==0.5.0
python-decouple==3.3
psycopg2==2.8.6